Finalize moving of bpf_filter.c. [skip ci]
As Blake Easley points it out in GitHub bug report #1166, before commit
08a5870 in 2018 bpf_filter.c was a symlink to bpf/net/bpf_filter.c, but
now it is a regular file, so it is wrong to list it in .gitignore.  Fix
that and reflect the file path change in DOS makefiles.

(backported from commit bbb12dc7568694271e33e658a3d99e579e5df08c)