Cirris CI: remove Windows task. [skip ci]
There are CI platforms that are useful for doing Windows CI using Visual
Studio.  It appears that Cirrus CI is not one of them, as they don't
appear to have images with VS pre-installed, and it does appear that
using Chocolatey to install it takes a long time.

(This was first removed from the main branch, after a failed attempt to
try to get it to work; the failure was because it requires the
aforementioned long slow installation of Visual Studio.)