Solaris: Fix AF_LINK handling on illumos.
The only supported OSes where get_sa_len() substitutes SA_LEN() are
Linux, illumos and Solaris 11.  On Linux the function handles the case
of AF_PACKET correctly, on Solaris 11 getifaddrs() does not return any
AF_LINK addresses, but on illumos the AF_LINK case defaults and
eventually dup_sockaddr() copies only the first sizeof(struct sockaddr)
bytes of struct sockaddr_dl, which is 16 bytes out of 252.

Right now on illumos because of a bug [1] .sdl_nlen is always set to 0,
which leaves 8 bytes for .sdl_data, which suffices for Ethernet and
Fibre Channel and the problem does not manifest in an obvious way.  I do
not have a setup to prove it would manifest for 20-byte InfiniBand
addresses, but as soon as the illumos bug is fixed, it will certainly
manifest for Ethernet, so let's fix it now.

While at it, add comments to fad-getad.c, fad-gifc.c and fad-glifc.c to
clarify which code paths apply to which OSes.

1: https://www.illumos.org/issues/16384
(cherry picked from commit c2454e001f54c59c7e847617e7e1b40148fd0756)